Remove search & first workspace page
Per design direction from Joana. On the Search page:

> It's quite discoverable because the search bar is there

On the workspace page:

> Too abstract

We keep the page that explains how to use gestures to switch between
workspaces.

The animations on the two touchpad gesture pages are magicked in via
CSS, which is based on the order of pages, so adjust those two magic
numbers for the removal of two pages.