Revert "welcome: Add link label to release notes"
This reverts commit 9d32a8d5ed0e7aa037f299d4650269c369458395.

This link points to the Endless OS 5 release notes, and hardcodes the
string “Endless OS 5”.

We added this link because the changes between Endless OS 4 and 5 were
so dramatic.

The tour is offered when the user logs in and org.gnome.shell
welcome-dialog-last-shown-version is set to a value lower than 40.beta.
Its default value is empty string, so in practice this means the tour i
offered in two cases:

1. New user account
2. Upgrade from Endless OS 4.

It is vanishingly unlikely that a user upgrading from Endless OS 4 will
be offered the tour for the first time in Endless OS 6. They would have
had to have performed the 4 → 5 upgrade, then, without logging into
their user account, upgraded from 5 to 6. So there is no need to tell
them about the 4 → 5 transition.

This leaves new users, who similarly do not need to be told about the 4
→ 5 transition.